Move scene examples into a better folder structure
