+++ examples/samples/scene/2d/index.js
@@ -1,0 +1,5 @@
+const basic = new URL('./basic.js', import.meta.url)
+
+export default {
+  'basic': basic
+}
